% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{river}
\alias{river}
\title{River data}
\format{
A data frame with 2 rows (\code{length} and \code{discharge}) and 8 variables:
\describe{
\item{\code{Missouri}}{ The Missouri River}
\item{\code{Mississippi}}{ The Mississippi River}
\item{\code{Yukon}}{ The Yukon River}
\item{\code{Colorado}}{ The Colorado River}
\item{\code{Arkansas}}{ The Arkansas River}
\item{\code{Columbia}}{ The Columbia River}
\item{\code{Red}}{ The Red River}
\item{\code{Canadian}}{ The Canadian River}
}
}
\usage{
river
}
\description{
A data frame containing length and discharge of several rivers
}
\keyword{datasets}
