% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/river_means.R
\name{river_means}
\alias{river_means}
\title{Means of river lengths and discharges}
\usage{
river_means(data, pattern)
}
\arguments{
\item{data}{A data frame with two rows. The first row indicates river length and
second row indicates river discharge. The columns of data indicate river names.}

\item{pattern}{A pattern by which to include only particular rivers}
}
\value{
The mean river length and mean river discharge for the desired rivers
}
\description{
Means of river lengths and discharges
}
\examples{
data("river")
river_means(river,"Mi")
}
