% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/river_stats.R
\name{river_stats}
\alias{river_stats}
\title{Summary statistics of river lengths and discharges}
\usage{
river_stats(data, pattern, FUN, ...)
}
\arguments{
\item{data}{A data frame with two rows. The first row indicates river length and
second row indicates river discharge. The columns of data indicate river names.}

\item{pattern}{A pattern by which to include only particular rivers}

\item{FUN}{A function to summarize the rivers by}

\item{...}{Additional arguments to \code{apply()}}
}
\value{
The summarized river length and mean river discharge for the desired rivers
}
\description{
Summary statistics of river lengths and discharges
}
\examples{
data("river")
river_stats(river, "Mi|C", mean, trim = 0.5)
}
